@@ -5,33 +5,34 @@ publish_to: none
 version: 1.10.2+27
 
 environment:
-  sdk: '>=2.17.0 <3.0.0'
+  sdk: '>=3.0.0 <4.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
 
+  collection: any
   flutter_launcher_icons: ^0.13.1
-  flutter_hooks: ^0.18.0
-  hooks_riverpod: ^2.3.6
-  riverpod_annotation: ^2.1.1
-  http: ^0.13.5
-  latlong2: ^0.8.0
-  csv: ^5.0.1
+  flutter_hooks: ^0.18.6
+  hooks_riverpod: ^2.4.6
+  riverpod_annotation: ^2.3.1
+  http: ^0.13.6
+  latlong2: ^0.8.2
+  csv: ^5.1.1
   intl: ^0.18.1
   flutter_map: ^3.0.0
   flutter_map_marker_cluster: ^1.0.1
   fl_chart: ^0.63.0
-  sqflite: ^2.2.0+3
-  shared_preferences: ^2.0.15
-  url_launcher: ^6.1.6
+  sqflite: ^2.3.0
+  shared_preferences: ^2.2.2
+  url_launcher: ^6.2.1
   package_info: ^2.0.2
   share: ^2.0.4
   webview_flutter: ^3.0.4
-  flutter_sticky_header: ^0.6.4
-  smooth_page_indicator: ^1.0.0+2
-  location: ^4.4.0
-  logging: ^1.1.1
+  flutter_sticky_header: ^0.6.5
+  smooth_page_indicator: ^1.1.0
+  location: ^5.0.3
+  logging: ^1.2.0
   # dart_rss:
   #   git:
   #     url: https://github.com/Chonli/webfeed.git
@@ -44,6 +45,9 @@ dev_dependencies:
   custom_lint: ^0.4.0
   build_runner: ^2.4.6
   riverpod_generator: ^2.2.3
+
+dependency_overrides:
+  watcher: 1.1.0
 
 flutter:
   uses-material-design: true
