@@ -6,124 +6,6 @@ part of 'preferences.dart';
 // RiverpodGenerator
 // **************************************************************************
 
-String _$themeModeSettingsHash() => r'4a73ee558c77ebf4690daaeec9b6c01d8b48f888';
-
-/// See also [ThemeModeSettings].
-@ProviderFor(ThemeModeSettings)
-final themeModeSettingsProvider =
-    AutoDisposeNotifierProvider<ThemeModeSettings, ThemeMode>.internal(
-  ThemeModeSettings.new,
-  name: r'themeModeSettingsProvider',
-  debugGetCreateSourceHash: const bool.fromEnvironment('dart.vm.product')
-      ? null
-      : _$themeModeSettingsHash,
-  dependencies: null,
-  allTransitiveDependencies: null,
-);
-
-typedef _$ThemeModeSettings = AutoDisposeNotifier<ThemeMode>;
-String _$lastStationDataSettingsHash() =>
-    r'd7d47c56e01106486a121b0fe566bb6c2cc73290';
-
-/// See also [LastStationDataSettings].
-@ProviderFor(LastStationDataSettings)
-final lastStationDataSettingsProvider =
-    AutoDisposeNotifierProvider<LastStationDataSettings, DateTime>.internal(
-  LastStationDataSettings.new,
-  name: r'lastStationDataSettingsProvider',
-  debugGetCreateSourceHash: const bool.fromEnvironment('dart.vm.product')
-      ? null
-      : _$lastStationDataSettingsHash,
-  dependencies: null,
-  allTransitiveDependencies: null,
-);
-
-typedef _$LastStationDataSettings = AutoDisposeNotifier<DateTime>;
-String _$lastStationSettingsHash() =>
-    r'24145f984b8c9299800e8e1d8f933167716d5020';
-
-/// See also [LastStationSettings].
-@ProviderFor(LastStationSettings)
-final lastStationSettingsProvider =
-    AutoDisposeNotifierProvider<LastStationSettings, DateTime>.internal(
-  LastStationSettings.new,
-  name: r'lastStationSettingsProvider',
-  debugGetCreateSourceHash: const bool.fromEnvironment('dart.vm.product')
-      ? null
-      : _$lastStationSettingsHash,
-  dependencies: null,
-  allTransitiveDependencies: null,
-);
-
-typedef _$LastStationSettings = AutoDisposeNotifier<DateTime>;
-String _$favoritesStationSettingsHash() =>
-    r'c8e34b1b1b02d208ce45e4a9ae11a69183473ef8';
-
-/// See also [FavoritesStationSettings].
-@ProviderFor(FavoritesStationSettings)
-final favoritesStationSettingsProvider = AutoDisposeNotifierProvider<
-    FavoritesStationSettings, List<String>>.internal(
-  FavoritesStationSettings.new,
-  name: r'favoritesStationSettingsProvider',
-  debugGetCreateSourceHash: const bool.fromEnvironment('dart.vm.product')
-      ? null
-      : _$favoritesStationSettingsHash,
-  dependencies: null,
-  allTransitiveDependencies: null,
-);
-
-typedef _$FavoritesStationSettings = AutoDisposeNotifier<List<String>>;
-String _$favoritesBERASettingsHash() =>
-    r'a1be8238d9537ce20744a54bb7aa830fd34234e1';
-
-/// See also [FavoritesBERASettings].
-@ProviderFor(FavoritesBERASettings)
-final favoritesBERASettingsProvider =
-    AutoDisposeNotifierProvider<FavoritesBERASettings, List<String>>.internal(
-  FavoritesBERASettings.new,
-  name: r'favoritesBERASettingsProvider',
-  debugGetCreateSourceHash: const bool.fromEnvironment('dart.vm.product')
-      ? null
-      : _$favoritesBERASettingsHash,
-  dependencies: null,
-  allTransitiveDependencies: null,
-);
-
-typedef _$FavoritesBERASettings = AutoDisposeNotifier<List<String>>;
-String _$showNoDataStationSettingsHash() =>
-    r'0882bf1c932d223fdd1c9e036a7ae80e5b48401d';
-
-/// See also [ShowNoDataStationSettings].
-@ProviderFor(ShowNoDataStationSettings)
-final showNoDataStationSettingsProvider =
-    AutoDisposeNotifierProvider<ShowNoDataStationSettings, bool>.internal(
-  ShowNoDataStationSettings.new,
-  name: r'showNoDataStationSettingsProvider',
-  debugGetCreateSourceHash: const bool.fromEnvironment('dart.vm.product')
-      ? null
-      : _$showNoDataStationSettingsHash,
-  dependencies: null,
-  allTransitiveDependencies: null,
-);
-
-typedef _$ShowNoDataStationSettings = AutoDisposeNotifier<bool>;
-String _$showClusterLayerSettingsHash() =>
-    r'8488f45b29101c368216b09f273ade082e62d844';
-
-/// See also [ShowClusterLayerSettings].
-@ProviderFor(ShowClusterLayerSettings)
-final showClusterLayerSettingsProvider =
-    AutoDisposeNotifierProvider<ShowClusterLayerSettings, bool>.internal(
-  ShowClusterLayerSettings.new,
-  name: r'showClusterLayerSettingsProvider',
-  debugGetCreateSourceHash: const bool.fromEnvironment('dart.vm.product')
-      ? null
-      : _$showClusterLayerSettingsHash,
-  dependencies: null,
-  allTransitiveDependencies: null,
-);
-
-typedef _$ShowClusterLayerSettings = AutoDisposeNotifier<bool>;
 String _$sharedPreferencesHash() => r'3a9f8412df34c1653d08100c9826aa2125b80f7f';
 
 /// See also [sharedPreferences].
@@ -139,4 +21,122 @@ final sharedPreferencesProvider = Provider<SharedPreferences>.internal(
 );
 
 typedef SharedPreferencesRef = ProviderRef<SharedPreferences>;
+String _$themeModeSettingsHash() => r'106550861a9acf66c527c4f3f77fe9c7c850ac47';
+
+/// See also [ThemeModeSettings].
+@ProviderFor(ThemeModeSettings)
+final themeModeSettingsProvider =
+    NotifierProvider<ThemeModeSettings, ThemeMode>.internal(
+  ThemeModeSettings.new,
+  name: r'themeModeSettingsProvider',
+  debugGetCreateSourceHash: const bool.fromEnvironment('dart.vm.product')
+      ? null
+      : _$themeModeSettingsHash,
+  dependencies: null,
+  allTransitiveDependencies: null,
+);
+
+typedef _$ThemeModeSettings = Notifier<ThemeMode>;
+String _$lastStationDataSettingsHash() =>
+    r'6cfdbd488e1416750f361235142136088aa8fd54';
+
+/// See also [LastStationDataSettings].
+@ProviderFor(LastStationDataSettings)
+final lastStationDataSettingsProvider =
+    NotifierProvider<LastStationDataSettings, DateTime>.internal(
+  LastStationDataSettings.new,
+  name: r'lastStationDataSettingsProvider',
+  debugGetCreateSourceHash: const bool.fromEnvironment('dart.vm.product')
+      ? null
+      : _$lastStationDataSettingsHash,
+  dependencies: null,
+  allTransitiveDependencies: null,
+);
+
+typedef _$LastStationDataSettings = Notifier<DateTime>;
+String _$lastStationSettingsHash() =>
+    r'106a627e5e1a04714b24b00973c082ed155258c1';
+
+/// See also [LastStationSettings].
+@ProviderFor(LastStationSettings)
+final lastStationSettingsProvider =
+    NotifierProvider<LastStationSettings, DateTime>.internal(
+  LastStationSettings.new,
+  name: r'lastStationSettingsProvider',
+  debugGetCreateSourceHash: const bool.fromEnvironment('dart.vm.product')
+      ? null
+      : _$lastStationSettingsHash,
+  dependencies: null,
+  allTransitiveDependencies: null,
+);
+
+typedef _$LastStationSettings = Notifier<DateTime>;
+String _$favoritesStationSettingsHash() =>
+    r'1558bd39ac145e4dad63815a1427b95068c6c379';
+
+/// See also [FavoritesStationSettings].
+@ProviderFor(FavoritesStationSettings)
+final favoritesStationSettingsProvider =
+    NotifierProvider<FavoritesStationSettings, List<String>>.internal(
+  FavoritesStationSettings.new,
+  name: r'favoritesStationSettingsProvider',
+  debugGetCreateSourceHash: const bool.fromEnvironment('dart.vm.product')
+      ? null
+      : _$favoritesStationSettingsHash,
+  dependencies: null,
+  allTransitiveDependencies: null,
+);
+
+typedef _$FavoritesStationSettings = Notifier<List<String>>;
+String _$favoritesBERASettingsHash() =>
+    r'fae0e555c2ed965494d055085abe2ded4603b9f4';
+
+/// See also [FavoritesBERASettings].
+@ProviderFor(FavoritesBERASettings)
+final favoritesBERASettingsProvider =
+    NotifierProvider<FavoritesBERASettings, List<String>>.internal(
+  FavoritesBERASettings.new,
+  name: r'favoritesBERASettingsProvider',
+  debugGetCreateSourceHash: const bool.fromEnvironment('dart.vm.product')
+      ? null
+      : _$favoritesBERASettingsHash,
+  dependencies: null,
+  allTransitiveDependencies: null,
+);
+
+typedef _$FavoritesBERASettings = Notifier<List<String>>;
+String _$showNoDataStationSettingsHash() =>
+    r'a8bc73ea9a30a056c0d89820ba8556cd972446e3';
+
+/// See also [ShowNoDataStationSettings].
+@ProviderFor(ShowNoDataStationSettings)
+final showNoDataStationSettingsProvider =
+    NotifierProvider<ShowNoDataStationSettings, bool>.internal(
+  ShowNoDataStationSettings.new,
+  name: r'showNoDataStationSettingsProvider',
+  debugGetCreateSourceHash: const bool.fromEnvironment('dart.vm.product')
+      ? null
+      : _$showNoDataStationSettingsHash,
+  dependencies: null,
+  allTransitiveDependencies: null,
+);
+
+typedef _$ShowNoDataStationSettings = Notifier<bool>;
+String _$showClusterLayerSettingsHash() =>
+    r'd6a4cd5ea6a4babe575c094592ec120df6c66d57';
+
+/// See also [ShowClusterLayerSettings].
+@ProviderFor(ShowClusterLayerSettings)
+final showClusterLayerSettingsProvider =
+    NotifierProvider<ShowClusterLayerSettings, bool>.internal(
+  ShowClusterLayerSettings.new,
+  name: r'showClusterLayerSettingsProvider',
+  debugGetCreateSourceHash: const bool.fromEnvironment('dart.vm.product')
+      ? null
+      : _$showClusterLayerSettingsHash,
+  dependencies: null,
+  allTransitiveDependencies: null,
+);
+
+typedef _$ShowClusterLayerSettings = Notifier<bool>;
 // ignore_for_file: unnecessary_raw_strings, subtype_of_sealed_class, invalid_use_of_internal_member, do_not_use_environment, prefer_const_constructors, public_member_api_docs, avoid_private_typedef_functions
