@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 import 'package:snow_weather_info/core/theme/app_theme.dart';
-import 'package:snow_weather_info/data/data_notifier.dart';
 import 'package:snow_weather_info/data/sources/preferences.dart';
-import 'package:snow_weather_info/ui/home_page.dart';
+import 'package:snow_weather_info/modules/home/home_page.dart';
+import 'package:snow_weather_info/provider/station_data.dart';
 
 Future<void> main() async {
   WidgetsFlutterBinding.ensureInitialized();
@@ -27,43 +27,6 @@ class MyApp extends ConsumerWidget {
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
-    /*return MultiProvider(
-      providers: [
-        Provider<DatabaseHelper>(create: (_) => DatabaseHelper()),
-        Provider<AvalancheAPI>(create: (_) => AvalancheAPI()),
-        Provider<Preferences>(create: (_) => Preferences(preferences)),
-        Provider<StationAPI>(create: (_) => StationAPI()),
-        ProxyProvider0<StationsRepository>(
-          update: (context, old) => StationsRepository.update(
-            context.watch<StationAPI>(),
-            context.watch<DatabaseHelper>(),
-            context.watch<Preferences>(),
-            old,
-          ),
-        ),
-        ChangeNotifierProxyProvider0<PreferenceNotifier>(
-          create: (_) => PreferenceNotifier(),
-          update: (context, old) => old!
-            ..preferences = context.watch<Preferences>()
-            ..init(),
-        ),
-        ChangeNotifierProvider<LocationNotifier>(
-          create: (_) => LocationNotifier(),
-        ),
-        ChangeNotifierProxyProvider0<DataNotifier>(
-          create: (_) => DataNotifier(),
-          update: (context, old) => old!
-            ..avalancheAPI = context.watch<AvalancheAPI>()
-            ..stationAPI = context.watch<StationAPI>()
-            ..preferences = context.watch<Preferences>()
-            ..databaseHelper = context.watch<DatabaseHelper>()
-            ..stationsRepository = context.watch<StationsRepository>()
-            ..initData(),
-        ),
-      ],
-      child: Consumer<PreferenceNotifier>(
-        builder: (context, PreferenceNotifier notifier, child) {*/
-
     return MaterialApp(
       title: _title,
       theme: AppTheme.ligthTheme,
@@ -84,7 +47,7 @@ class MyHomePage extends ConsumerWidget {
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
-    final loading = ref.watch(dataNotifier.select((n) => n.loading));
+    final loading = ref.watch(stationDataProvider.select((n) => n.isLoading));
 
     return loading ? const _LoadingPage() : HomePage(title: title);
   }
