@@ -1,25 +1,25 @@
 // GENERATED CODE - DO NOT MODIFY BY HAND
 
-part of 'stations.dart';
+part of 'avalanche.dart';
 
 // **************************************************************************
 // RiverpodGenerator
 // **************************************************************************
 
-String _$stationRepositoryHash() => r'82c845c882f46e17b6a8f091c1a6a69eedebd659';
+String _$avalancheFeedHash() => r'03384d5aeecf88d8b40d9b76c9b9476ff259f0c0';
 
-/// See also [StationRepository].
-@ProviderFor(StationRepository)
-final stationRepositoryProvider =
-    AutoDisposeAsyncNotifierProvider<StationRepository, List<Station>>.internal(
-  StationRepository.new,
-  name: r'stationRepositoryProvider',
+/// See also [AvalancheFeed].
+@ProviderFor(AvalancheFeed)
+final avalancheFeedProvider =
+    AsyncNotifierProvider<AvalancheFeed, int>.internal(
+  AvalancheFeed.new,
+  name: r'avalancheFeedProvider',
   debugGetCreateSourceHash: const bool.fromEnvironment('dart.vm.product')
       ? null
-      : _$stationRepositoryHash,
+      : _$avalancheFeedHash,
   dependencies: null,
   allTransitiveDependencies: null,
 );
 
-typedef _$StationRepository = AutoDisposeAsyncNotifier<List<Station>>;
+typedef _$AvalancheFeed = AsyncNotifier<int>;
 // ignore_for_file: unnecessary_raw_strings, subtype_of_sealed_class, invalid_use_of_internal_member, do_not_use_environment, prefer_const_constructors, public_member_api_docs, avoid_private_typedef_functions
